[refactor] Extract to separate class files
