+++ exist-core/src/main/java/org/exist/storage/serializers/XmlSerializerPoolObjectFactory.java
@@ -1,0 +1,54 @@
+/*
+ * eXist-db Open Source Native XML Database
+ * Copyright (C) 2001 The eXist-db Authors
+ *
+ * info@exist-db.org
+ * http://www.exist-db.org
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 2.1 of the License, or (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
+ */
+package org.exist.storage.serializers;
+
+import org.apache.commons.pool2.BasePooledObjectFactory;
+import org.apache.commons.pool2.PooledObject;
+import org.apache.commons.pool2.impl.DefaultPooledObject;
+import org.exist.storage.DBBroker;
+import org.exist.util.Configuration;
+
+public class XmlSerializerPoolObjectFactory extends BasePooledObjectFactory<Serializer> {
+
+    private final DBBroker broker;
+    private final Configuration config;
+
+    public XmlSerializerPoolObjectFactory(final DBBroker broker, final Configuration config) {
+        this.broker = broker;
+        this.config = config;
+    }
+
+    @Override
+    public Serializer create() {
+        return new NativeSerializer(broker, config);
+    }
+
+    @Override
+    public PooledObject wrap(final Serializer obj) {
+        return new DefaultPooledObject<>(obj);
+    }
+
+    @Override
+    public void passivateObject(final PooledObject<Serializer> p) {
+        p.getObject().reset();
+    }
+}
